refactor: separate dsc library from dsc binary
The motivation for this change is so that I can import the engine
of docker-source-checksum as a function in some other binary,
without introducing changes to current docker-source-checksum users.
